Fix a config save issue.
When the font size was changed, an invalid config would be written.
If the user did not press the Save button to overwrite this invalid
config, then subsequent loads of the config page were unable to
overwrite due to loading the invalid config.  This patch fixes the
issue with the font size so that it is not saved until the save
button is clicked and it also provides a workaround for users who
already have the invalid config file containing only the word 'null'.